moved news to new directories
