+++ dook/api/news/errors.py
@@ -1,0 +1,6 @@
+MULTIPLE_OPINIONS_FROM_ONE_USER_FOR_ONE_NEWS_ERROR = (
+    "User can have only one opinion per one news"
+)
+EVENT_MISSING_VERDICT_ATTRIBUTE_ERROR = (
+    "_verdict attribute is missing. Call is_with_verdict()"
+)
