+++ dook/core/news/constants.py
