+++ dook/core/news/events.py
@@ -1,9 +1,10 @@
+from dook.api.news.errors import EVENT_MISSING_VERDICT_ATTRIBUTE_ERROR
 from dook.core.events.services import ModelEventService
 
 
 class NewsEvents(ModelEventService):
     def new_verdict(self, *args, **kwargs):
         if not hasattr(self.obj, "_verdict"):
-            raise AttributeError("_verdict attribute is missing. Call is_with_verdict()")
+            raise AttributeError(EVENT_MISSING_VERDICT_ATTRIBUTE_ERROR)
 
         self._send("news_new_verdict", *args, **kwargs)
