+++ dook/core/news/managers.py
