+++ dook/core/news/models.py
@@ -5,9 +5,9 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.db import IntegrityError, models
 
 from dook.core.events.mixins import ModelEventMixin
-from dook.news.constants import VerdictType
-from dook.news.events import NewsEvents
-from dook.news.managers import NewsManager, NewsSensitiveKeywordsManager
+from dook.core.news.constants import VerdictType
+from dook.core.news.events import NewsEvents
+from dook.core.news.managers import NewsManager, NewsSensitiveKeywordsManager
 from dook.users.constants import UserRoleType
 from dook.users.models import User
